python311-services/pyproject.toml
[tool.poetry]
name = "python311-services"
version = "0.1.0"
description = "Python 3.11 compatibility services for Hybrid AI Council - handles legacy ML dependencies"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "voice"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# Core web framework for microservice
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
aiohttp = "^3.9.0"
pydantic = "^2.0.0"

# Voice Processing Dependencies (All require Python <3.12)
coqui-tts = "^0.27.0"              # Coqui XTTS v2 (NEW - main TTS engine)
torch = "^2.7.1"                  # Core PyTorch
torchaudio = "^2.7.1"             # Audio processing  
transformers = "^4.54.1"          # HuggingFace models
nemo-toolkit = "^2.4.0"           # NVIDIA NeMo (Parakeet STT)
# silero-vad removed - Parakeet STT handles non-speech segments naturally
pytorch-lightning = "^2.5.2"      # Training framework
lightning = "^2.5.2"              # Lightning umbrella
lhotse = "^1.30.3"                # Audio dataset toolkit
jiwer = "^4.0.0"                  # Word error rate calculation

# Core utilities
structlog = "^23.0.0"              # Structured logging
python-dotenv = "^1.0.0"           # Environment management
httpx = "^0.25.0"                  # HTTP client

# Audio processing
numpy = "^1.25.0"                  # Numerical computing
librosa = "^0.10.0"                # Audio analysis (if needed)

# Note: faster-whisper removed - using NeMo Parakeet STT only

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true